% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv_airplane.R
\name{rv_create_airplane}
\alias{rv_create_airplane}
\title{Create risk airplane}
\usage{
rv_create_airplane(
  risk_ratio = 0.1,
  risk_places = NULL,
  risk_vector = NULL,
  compact = FALSE,
  risk_names = "Risk",
  fill = c("coral1", "darkcyan", "darkgoldenrod1", "darkseagreen1", "lightskyblue1"),
  na_colour = "grey80",
  rows = 33,
  seats_per_row = 6,
  title = NULL,
  font_family = "sans",
  font_size = 20,
  font_family_seats = "mono",
  seat_size_ratio = 1,
  legend_position = "none",
  coord_ratio = 0.8
)
}
\arguments{
\item{risk_places}{Number of risk places, defaults to NULL. If given, takes precedence over ratio.}

\item{fill}{A vector of colours.}

\item{rows}{Number of rows, defaults to 33 (found in commercial airplans, and with 6 seats per rows it adds up to 198 seats in total).}

\item{legend_position}{Defaults to "none".}

\item{coord_ratio}{Numeric, defaults to 0.5, untested with anything else. Y/X ratio.}
}
\value{

}
\description{
Create risk airplane
}
