% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv_airplane.R
\name{rv_create_airplane_animation}
\alias{rv_create_airplane_animation}
\title{Create animated version with multiple scenarios}
\usage{
rv_create_airplane_animation(
  risk_ratio = c(`Scenario A` = 0.01, `Scenario B` = 0.1, `Scenario C` = 0.3, `Scenario
    D` = 0.7),
  transition_length = 0.5,
  state_length = 1,
  risk_places = NULL,
  risk_vector = NULL,
  risk_name = "Risk",
  fill = c("coral1", "darkgoldenrod1", "darkseagreen1", "lightskyblue1"),
  na_colour = "grey80",
  rows = 33,
  seats_per_row = 6,
  title = NULL,
  font_family = "sans",
  font_family_seats = "mono",
  legend_position = "none",
  ncol = NULL,
  nrow = NULL,
  guides = NULL
)
}
\arguments{
\item{risk_ratio}{}

\item{transition_length}{}

\item{state_length}{}

\item{risk_places}{}

\item{risk_vector}{}

\item{risk_name}{}

\item{fill}{}

\item{na_colour}{}

\item{rows}{}

\item{title}{}

\item{font_family}{}

\item{font_family_seats}{}

\item{legend_position}{}

\item{ncol}{}

\item{nrow}{}

\item{guides}{}
}
\value{

}
\description{
Create animated version with multiple scenarios
}
