% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv_airplane.R
\name{rv_create_airplane_combo}
\alias{rv_create_airplane_combo}
\title{Spread the risk among more planes}
\usage{
rv_create_airplane_combo(
  risk_ratio = 0.1,
  number_of_planes = 2,
  risk_places = NULL,
  risk_vector = NULL,
  compact = FALSE,
  risk_names = NULL,
  fill = c("coral1", "darkcyan", "darkgoldenrod1", "darkseagreen1", "lightskyblue1"),
  na_colour = "grey90",
  rows = 33,
  seats_per_row = 6,
  title = NULL,
  font_family = "sans",
  font_family_seats = "mono",
  legend_position = "none",
  ncol = NULL,
  nrow = NULL,
  guides = NULL
)
}
\arguments{
\item{seats_per_row}{Defaults to 6. Untested with anything different from 6.}

\item{legend_position}{}
}
\description{
Spread the risk among more planes
}
