% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv_airplane_utils.R
\name{rv_create_airplane_graph_compact}
\alias{rv_create_airplane_graph_compact}
\title{Create a ggplot based on seats typically generated with `rv_create_airplane_seats()`}
\usage{
rv_create_airplane_graph_compact(
  seats,
  rows = 33,
  fill = c("coral1", "darkcyan", "darkgoldenrod1", "darkseagreen1", "lightskyblue1"),
  na_colour = "grey80",
  levels = NULL,
  title = NULL,
  font_family = "sans",
  legend_position = "none"
)
}
\arguments{
\item{seats}{}

\item{rows}{}

\item{fill}{}

\item{na_colour}{}

\item{title}{}

\item{font_family}{}

\item{legend_position}{}
}
\value{

}
\description{
Create a ggplot based on seats typically generated with `rv_create_airplane_seats()`
}
