% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv_airplane_utils.R
\name{rv_create_airplane_seats}
\alias{rv_create_airplane_seats}
\title{Create a dataframe with seats location and names (used internally)}
\usage{
rv_create_airplane_seats(rows = 33, seats_per_row = 6)
}
\arguments{
\item{rows}{Numeric, defaults to 33}

\item{seats_per_row}{Numeric, defaults to 6. Untested with anything different from 6.}
}
\value{

}
\description{
Create a dataframe with seats location and names (used internally)
}
