% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv_arena.R
\name{rv_create_arena}
\alias{rv_create_arena}
\title{Create an arena with a given ratio of places highlighted}
\usage{
rv_create_arena(
  ratio,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  bg = "white",
  quality = "low",
  ggplot = FALSE
)
}
\arguments{
\item{ratio}{Numeric, between 0 and 1, required. Share of seats to colour.}

\item{bg}{Character vector of lenght 1, defines the background colour of the plot. Defaults to "white".}

\item{ggplot}{Logical, defaults to FALSE. If TRUE, returns a `ggplot` object. If FALSE, a `magick` object.}
}
\value{

}
\description{
Create an arena with a given ratio of places highlighted
}
\examples{

rv_create_arena(ratio = 0.9)

rv_create_arena(ratio = 0.1)

}
