% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv_create_risk_vector.R
\name{rv_create_risk_vector}
\alias{rv_create_risk_vector}
\title{Create risk vector based on ratio, number of places, or data frame}
\usage{
rv_create_risk_vector(
  risk_ratio,
  total_places,
  risk_places = NULL,
  risk_name = "Risk"
)
}
\arguments{
\item{risk_ratio}{A numeic vector or a data frame with two columns, `Risk` and `Ratio`}

\item{total_places}{Numeric. Gives the total length of the output vector.}

\item{risk_places}{A numeric vector or a data frame with two columns, `Risk` and `Places`}

\item{risk_name}{If a numeric vector given in previous parameters, defines the name of the risk.}
}
\value{
A character vector with length equal to total_places.
}
\description{
Create risk vector based on ratio, number of places, or data frame
}
\examples{


rv_create_risk_vector(risk_ratio = 0.1,
                      total_places = 100)
table(
  rv_create_risk_vector(risk_ratio = 0.1,
                        total_places = 100) 
)

ratio <- tibble::tribble(~Risk, ~Ratio,
                         "Hospitalization", 0.3,
                         "Death", 0.1)

rv_create_risk_vector(risk_ratio = ratio,
                      total_places = 100)

table(
  rv_create_risk_vector(risk_ratio = ratio,
                        total_places = 100)
)


}
