% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv_explorer.R
\name{rv_explorer}
\alias{rv_explorer}
\title{Run the Shiny Application}
\usage{
rv_explorer(...)
}
\arguments{
\item{...}{A series of options to be used inside the app.}
}
\description{
Run the Shiny Application
}
