% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv_magick.R
\name{rv_img}
\alias{rv_img}
\title{Converts output to an image object in standard formats.}
\usage{
rv_img(plot, quality = "low", bg = "white", stack = FALSE)
}
\arguments{
\item{plot}{A plot, typically generated with e.g. `rv_create_arena()`}

\item{quality}{One of either "low" (output 1024px width), "medium" (output 1512px width), or "high" (output 2048px width). Defaults to "low".}

\item{bg}{Background colour.}

\item{stack}{Logical, defaults to FALSE. If TRUE, graphs are placed left-to-right, if FALSE top-to-bottom.}
}
\description{
Converts output to an image object in standard formats.
}
\examples{

plot <- rv_create_arena(ratio = 0.7)
rv_img(plot, quality = "medium")

}
