% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv_create_airplane_compact.R
\name{rv_create_airplane_compact}
\alias{rv_create_airplane_compact}
\title{Create a compact airplane}
\usage{
rv_create_airplane_compact(
  risk_ratio,
  risk_vector = NULL,
  risk_names = NULL,
  rows = 33,
  seats_per_row = 6,
  fill = c("coral1", "darkcyan", "darkgoldenrod1", "darkseagreen1", "lightskyblue1"),
  na_colour = "grey80",
  title = NULL,
  font_family = "sans",
  legend_position = "none"
)
}
\arguments{
\item{risk_vector}{A character vector of length equal to the number of seats (rows*seats_per_row)}

\item{rows}{}

\item{fill}{A character vector of colours.}

\item{na_colour}{}

\item{title}{}

\item{font_family}{}

\item{legend_position}{}
}
\value{

}
\description{
Create a compact airplane
}
